synced dev-python/pycairo with portage tree

+++ dev-python/pycairo/pycairo-1.8.6.ebuild
@@ -1,8 +1,9 @@
 # Copyright 1999-2009 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/dev-python/pycairo/pycairo-1.8.6.ebuild,v 1.1 2009/07/09 03:24:42 arfrever Exp $
+# $Header: /var/cvsroot/gentoo-x86/dev-python/pycairo/pycairo-1.8.6.ebuild,v 1.3 2009/08/08 20:36:20 arfrever Exp $
 
-EAPI=2
+EAPI="2"
+
 NEED_PYTHON="2.6"
 SUPPORT_PYTHON_ABIS="1"
 
@@ -15,33 +16,58 @@ SRC_URI="http://cairographics.org/releases/${P}.tar.gz"
 LICENSE="|| ( LGPL-2.1 MPL-1.1 )"
 SLOT="0"
 KEYWORDS="~alpha ~amd64 ~arm ~hppa ~ia64 ~mips ~ppc ~ppc64 ~sh ~sparc ~x86 ~x86-fbsd"
-IUSE="examples"
+IUSE="doc examples"
 
 RDEPEND=">=x11-libs/cairo-1.8.6[lib32?]"
 DEPEND="${RDEPEND}
-	dev-util/pkgconfig[lib32?]"
+	dev-util/pkgconfig[lib32?]
+	doc? ( dev-python/sphinx )"
+
+RESTRICT_PYTHON_ABIS="2.4 2.5 3*"
 
 PYTHON_MODNAME="cairo"
-DOCS="AUTHORS NEWS doc/*"
+DOCS="AUTHORS NEWS README"
 
 multilib-native_src_prepare_internal() {
-	# don't run py-compile.
+	# Don't run py-compile.
 	sed -i \
 		-e '/if test -n "$$dlist"; then/,/else :; fi/d' \
 		cairo/Makefile.in || die "sed in cairo/Makefile.in failed"
 }
 
+multilib-native_src_configure_internal() {
+	if use doc; then
+		econf
+	fi
+}
+
+multilib-native_src_compile_internal() {
+	distutils_src_compile
+
+	if use doc; then
+		emake html || die "emake html failed"
+	fi
+}
+
+src_test() {
+	testing() {
+		pushd test > /dev/null
+		PYTHONPATH="$(ls -d ../build-${PYTHON_ABI}/lib.*)" "$(PYTHON)" test.py || return 1
+		popd > /dev/null
+	}
+	python_execute_function testing
+}
+
 multilib-native_src_install_internal() {
 	distutils_src_install
+
+	if use doc; then
+		dohtml -r doc/.build/html/ || die "dohtml -r doc/.build/html/ failed"
+	fi
 
 	if use examples; then
 		insinto /usr/share/doc/${PF}/examples
 		doins -r examples/*
 		rm "${D}"/usr/share/doc/${PF}/examples/Makefile*
 	fi
-}
-
-src_test() {
-	cd test
-	PYTHONPATH="$(ls -d ${S}/build/lib.*)" "${python}" test.py || die "tests failed"
 }
